requirements.txt

# Core LLM & Agent Framework
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
langgraph==0.2.0
openai==1.50.0

# Vector Store
pinecone-client==5.0.0
pinecone-text==0.9.0          # for BM25 sparse embeddings

# API
fastapi==0.115.0
uvicorn[standard]==0.30.0
mangum==0.17.0                 # ASGI adapter for AWS Lambda

# Utils
python-dotenv==1.0.0
pydantic==2.9.0
httpx==0.27.0
tiktoken==0.7.0

# AWS
boto3==1.35.0
